Create `create.rs` file to handle the business logic
